feat: renamed Headerneu to Hero
